refactor: peer review discussion
