+++ packages/core/src/runtime/pipeline/index.ts
@@ -2,4 +2,4 @@ export * from "./agent";
 export * from "./operations";
 export * from "./templates";
 export * from "./types";
-export * from "./processor";
+export * from "./scheduler";
