+++ packages/core/src/runtime/pipeline/operations.ts
@@ -1,3 +1,4 @@
+// DEV NOTE: GET RID OF THIS ENTIRE FILE OR MOVE THE CONTENTS TO SOMEWHERE ELSE
 import { z } from "zod";
 
 import { ModelProvider, ModelRequestConfig } from "../providers/model";
